+++ tests/MailMimeParser/Parser/PartBuilderFactoryTest.php
@@ -3,6 +3,7 @@ namespace ZBateson\MailMimeParser\Parser;
 
 use LegacyPHPUnit\TestCase;
 use GuzzleHttp\Psr7;
+use GuzzleHttp\Psr7\StreamWrapper;
 
 /**
  * PartBuilderFactoryTest
@@ -23,8 +24,12 @@ class PartBuilderFactoryTest extends TestCase
 
     public function testNewPartBuilder()
     {
+        $hc = $this->getMockBuilder('ZBateson\MailMimeParser\Message\PartHeaderContainer')
+            ->disableOriginalConstructor()
+            ->getMock();
         $stream = Psr7\stream_for('test');
         $partBuilder = $this->instance->newPartBuilder(
+            $hc,
             $stream
         );
         $this->assertInstanceOf(
@@ -32,13 +37,23 @@ class PartBuilderFactoryTest extends TestCase
             $partBuilder
         );
 
+        $parent = $this->getMockBuilder('ZBateson\MailMimeParser\Parser\Proxy\ParserPartProxy')
+            ->disableOriginalConstructor()
+            ->setMethods([ 'getMessageResourceHandle' ])
+            ->getMockForAbstractClass();
+        $parent->expects($this->once())
+            ->method('getMessageResourceHandle')
+            ->willReturn(StreamWrapper::getResource($stream));
+
         $childPartBuilder = $this->instance->newChildPartBuilder(
-            $partBuilder
+            $hc,
+            $parent
         );
         $this->assertInstanceOf(
             '\ZBateson\MailMimeParser\Parser\PartBuilder',
             $childPartBuilder
         );
+        $this->assertSame(0, $childPartBuilder->getStreamPartStartPos());
         $stream->close();
     }
 }
